+++ packages/client/observable/operators/index.js
@@ -1,0 +1,1 @@
+module.exports = require('../../dist/observable/operators/index');
