+++ packages/client/rollup.config.js
@@ -9,8 +9,8 @@ const config = getRollupConfig({
     'src/links/splitLink.ts',
     'src/links/transformerLink.ts',
     'src/links/wsLink.ts',
-    'src/rx/index.ts',
-    'src/rx/operators/index.ts',
+    'src/observable/index.ts',
+    'src/observable/operators/index.ts',
   ],
 });
