--- packages/client/rx/index.js
@@ -1,1 +1,0 @@
-module.exports = require('../dist/rx/index');
