--- packages/client/rx/operators/index.d.ts
@@ -1,1 +1,0 @@
-export * from '../../dist/rx/operators';
