+++ packages/client/src/links/dedupeLink.test.ts
@@ -1,7 +1,7 @@
 import { waitFor } from '@testing-library/dom';
 import { AnyRouter } from '@trpc/server';
 import { OperationLink } from '..';
-import { observable } from '../rx/observable';
+import { observable } from '../observable/observable';
 import { dedupeLink } from './dedupeLink';
 import { createChain } from './internals/createChain';
