+++ packages/client/src/links/retryLink.ts
@@ -1,6 +1,6 @@
 import { AnyRouter } from '@trpc/server';
-import { observable } from '../rx/observable';
-import { Unsubscribable } from '../rx/types';
+import { observable } from '../observable/observable';
+import { Unsubscribable } from '../observable/types';
 import { TRPCLink } from './types';
 
 export function retryLink<TRouter extends AnyRouter = AnyRouter>(opts: {
