+++ packages/client/src/links/splitLink.ts
@@ -1,5 +1,5 @@
 import { AnyRouter } from '@trpc/server';
-import { observable } from '../rx/observable';
+import { observable } from '../observable/observable';
 import { TRPCLink, Operation } from './types';
 import { createChain } from './internals/createChain';
