+++ packages/client/src/links/wsLink.ts
@@ -6,8 +6,8 @@ import {
   TRPCResponse,
 } from '@trpc/server/rpc';
 import { retryDelay } from '../internals/retryDelay';
-import { UnsubscribeFn } from '../rx';
-import { observable } from '../rx/observable';
+import { UnsubscribeFn } from '../observable';
+import { observable } from '../observable/observable';
 import { TRPCClientError } from '../TRPCClientError';
 import { Operation, OperationResultObserver, TRPCLink } from './types';
