+++ packages/client/src/observable/index.ts
@@ -1,2 +1,3 @@
 export * from './observable';
 export * from './types';
+export * from './operators';
