+++ packages/client/src/observable/operators/map.ts
