+++ packages/server/test/links.test.ts
@@ -4,14 +4,14 @@ import { createChain } from '@trpc/client/src/links/internals/createChain';
 import { LinkRuntime, OperationLink } from '@trpc/client/src/links/types';
 import AbortController from 'abort-controller';
 import fetch from 'node-fetch';
-import { observableToPromise } from '../../client/src/rx/util/observableToPromise';
+import { observableToPromise } from '../../client/src/observable/util/observableToPromise';
 import { z } from 'zod';
 import { createTRPCClient, TRPCClientError } from '../../client/src';
 import { httpBatchLink } from '../../client/src/links/httpBatchLink';
 import { httpLink } from '../../client/src/links/httpLink';
 import { loggerLink } from '../../client/src/links/loggerLink';
 import { retryLink } from '../../client/src/links/retryLink';
-import { observable } from '../../client/src/rx/observable';
+import { observable } from '../../client/src/observable/observable';
 import * as trpc from '../src';
 import { AnyRouter } from '../src';
 import { routerToServerAndClient } from './_testHelpers';
